+++ CHANGELOG.md
@@ -32,8 +32,8 @@
     * CBOR
 
 * details
-  * Revision 996
-    * finding github ctest segfault issue ...
+  * Revision 997
+    * github ci openssl 3.0.0 environment
   * Revision 994
     * [fixed] floating point, bignumber
     * [changed] self-signed certificates